pyproject.toml
[project]
name = "python-gemini-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.60.0",
    "python-dotenv>=1.2.1",
]
